+++ R/data.R
@@ -1,0 +1,19 @@
+#' AFL Tables Mapping: AFL Tables Naming Convention Mapping
+#'
+#' A dataframe used for mapping names of data columns when scraping match statistics using the `scrape_afltables_match()` function.
+#'
+#' @keywords internal
+"mapping_afltables"
+
+#' AFL Tables player mapping: AFL Tables Player Name Mapping
+#'
+#' A dataframe used for mapping player name and ID columns when scraping match statistics using the `scrape_afltables_match()` function.
+#'
+#' @keywords internal
+"player_mapping_afltables"
+
+#' AFL Tables Data Dictionary: AFL Tables Column Formats
+#'
+#' A dataframe containing the data types and formats of columns for AFL Tables data, used in the `check_and_convert()` function to ensure correct column types.
+#' @keywords internal
+"dictionary_afltables"