+++ R/helpers-afltables-playerstats.R
@@ -10,6 +10,33 @@
 #' @keywords internal
 #' @noRd
 scrape_afltables_match <- function(match_urls) {
+  # Function to extract a table based on some unique text it contains
+  extract_table <- function(page, unique_text) {
+    tables <- rvest::html_nodes(page, "table")
+    for (i in seq_along(tables)) {
+      table <- tables[[i]]
+      if (any(stringr::str_starts(rvest::html_text(table), unique_text))) {
+        links <- table %>% rvest::html_nodes("a")
+        urls <- links %>% rvest::html_attr("href")
+        filtered_urls <- urls %>% stringr::str_subset("/players/")
+        updated_urls <- stringr::str_replace(filtered_urls, "^\\.\\.\\/\\.\\.\\/", "https://afltables.com/afl/stats/")
+        table_df <- rvest::html_table(table) %>%
+          janitor::row_to_names(row_number = 1) %>%
+          dplyr::filter(!.data$Player %in% c("Rushed", "Totals", "Opposition"))
+
+        if (length(updated_urls) == nrow(table_df)) {
+          table_tot <- dplyr::bind_cols(table_df, url = updated_urls)
+        } else {
+          table_tot <- table_df
+          table_tot$url <- NA
+        }
+
+        return(table_tot)
+      }
+    }
+    return(NULL)
+  }
+
   scrape_afltables_data <- function(url) {
     # Read the webpage content
     page <- rvest::read_html(url)
@@ -18,30 +45,102 @@ scrape_afltables_match <- function(match_urls) {
     details <- rvest::html_node(page, "table") %>% rvest::html_table()
     team_names <- details$X2[2:3]
 
-    # Function to extract a table based on some unique text it contains
-    extract_table <- function(page, unique_text) {
-      tables <- rvest::html_nodes(page, "table")
-      for (i in seq_along(tables)) {
-        table <- tables[[i]]
-        if (any(stringr::str_starts(rvest::html_text(table), unique_text))) {
-          return(rvest::html_table(table) %>%
-            janitor::row_to_names(row_number = 1))
-        }
-      }
-      return(NULL)
-    }
-
     # Extract the match statistics tables for both teams
     team1_stats <-
       extract_table(page, paste0(team_names[1], " Match Statistics")) %>%
-      dplyr::mutate(Playing.for = team_names[1])
-    team2_stats <- extract_table(page, paste0(team_names[2], " Match Statistics")) %>%
-      dplyr::mutate(Playing.for = team_names[2])
+      dplyr::mutate(Playing.for = team_names[1]) %>%
+      dplyr::rename(dplyr::any_of(mapping_afltables)) %>%
+      dplyr::mutate(
+        Substitute = dplyr::case_when(
+          stringr::str_detect(.data$Jumper.No., "\u2191") ~ "On",
+          stringr::str_detect(.data$Jumper.No., "\u2193") ~ "Off",
+          TRUE ~ NA_character_
+        ),
+        Jumper.No. = stringr::str_remove_all(.data$Jumper.No., "[\u2191\u2193]") %>%
+          as.character() %>% stringr::str_trim()
+      ) %>%
+      check_and_convert(dictionary_afltables)
+
+    team2_stats <-
+      extract_table(page, paste0(team_names[2], " Match Statistics")) %>%
+      dplyr::mutate(Playing.for = team_names[2]) %>%
+      dplyr::rename(dplyr::any_of(mapping_afltables)) %>%
+      dplyr::mutate(
+        Substitute = dplyr::case_when(
+          stringr::str_detect(.data$Jumper.No., "\u2191") ~ "On",
+          stringr::str_detect(.data$Jumper.No., "\u2193") ~ "Off",
+          TRUE ~ NA_character_
+        ),
+        Jumper.No. = stringr::str_remove_all(.data$Jumper.No., "[\u2191\u2193]") %>%
+          as.character() %>% stringr::str_trim()
+      ) %>%
+      dplyr::filter(!.data$Player %in% c("Rushed", "Totals", "Opposition")) %>%
+      check_and_convert(dictionary_afltables)
+
+    team1_details <-
+      extract_table(page, paste0(team_names[1], " Player Details")) %>%
+      dplyr::mutate(Playing.for = team_names[1]) %>%
+      dplyr::rename(dplyr::any_of(mapping_afltables)) %>%
+      dplyr::mutate(
+        Age = suppressWarnings(convert_age_to_years(.data$Age)), # Convert age to decimal years
+        Career.Games = as.numeric(sub(" .*", "", .data$Career.Games)) # Extract the number of games before the bracket and convert to numeric
+      ) %>%
+      dplyr::select("Jumper.No.", "Player", "Age", "Career.Games")
+
+    team2_details <-
+      extract_table(page, paste0(team_names[2], " Player Details")) %>%
+      dplyr::mutate(Playing.for = team_names[2]) %>%
+      dplyr::rename(dplyr::any_of(mapping_afltables)) %>%
+      dplyr::mutate(
+        Age = suppressWarnings(convert_age_to_years(.data$Age)), # Convert age to decimal years
+        Career.Games = as.numeric(sub(" .*", "", .data$Career.Games)) # Extract the number of games before the bracket and convert to numeric
+      ) %>%
+      dplyr::select("Jumper.No.", "Player", "Age", "Career.Games")
+
+    # Ensure there's only one row with Jumper.No. equal to "C"
+    home_coach <- team1_details %>%
+      dplyr::filter(.data$Jumper.No. == "C") %>%
+      dplyr::pull(.data$Player)
+
+    away_coach <- team2_details %>%
+      dplyr::filter(.data$Jumper.No. == "C") %>%
+      dplyr::pull(.data$Player)
+
+    if (length(home_coach) != 1) {
+      home_coach <- NA
+    }
+    if (length(away_coach) != 1) {
+      away_coach <- NA
+    }
+
+    # Bind data together
+    home_total <-
+      team1_stats %>%
+      dplyr::left_join(team1_details,
+        by = c("Jumper.No.", "Player")
+      ) %>%
+      dplyr::mutate(
+        Coach = home_coach,
+        Home.Away = "Home"
+      )
+
+    away_total <-
+      team2_stats %>%
+      dplyr::left_join(team2_details,
+        by = c("Jumper.No.", "Player")
+      ) %>%
+      dplyr::mutate(
+        Coach = away_coach,
+        Home.Away = "Away"
+      )
+
+    tot_stats <-
+      dplyr::bind_rows(home_total, away_total) %>%
+      replace_na_conditionally()
 
     return(list(
-      details = details,
-      home_games = team1_stats,
-      away_games = team2_stats
+      game_details = details,
+      tot_stats = tot_stats
     ))
   }
   cli::cli_progress_step("Scraping AFL Tables")
@@ -50,15 +149,11 @@ scrape_afltables_match <- function(match_urls) {
 
   cli::cli_progress_step("Cleaning AFL Tables data")
 
-  details <- purrr::map(scraped_data, purrr::pluck, "details")
-  home_games <- purrr::map(scraped_data, purrr::pluck, "home_games")
-  away_games <- purrr::map(scraped_data, purrr::pluck, "away_games")
+  game_details <- purrr::map(scraped_data, purrr::pluck, "game_details")
+  games_tot <- purrr::map(scraped_data, purrr::pluck, "tot_stats")
 
-  games_tot <- home_games %>%
-    purrr::map2(.y = away_games, ~ dplyr::bind_rows(.x, .y))
-
-  att_lgl <- details %>%
-    purrr::map(~ any(stringr::str_detect(.x[2,], "Attendance")))
+  att_lgl <- game_details %>%
+    purrr::map(~ any(stringr::str_detect(.x[1, 2], "Attendance")))
 
   att_fn <- function(x) {
     if (x) {
@@ -71,22 +166,26 @@ scrape_afltables_match <- function(match_urls) {
   date_str <- att_lgl %>%
     purrr::map(att_fn)
 
-  args <- list(games_tot, details, date_str)
+  args <- list(games_tot, game_details, date_str)
 
   games_df <- args %>%
     purrr::pmap(~ dplyr::mutate(..1, Date = stringr::str_extract(..2[1, 2], ..3)))
 
   games_df <- games_df %>%
-    purrr::map2(.y = details, ~ dplyr::mutate(
+    purrr::map2(.y = game_details, ~ dplyr::mutate(
       .x,
       Round = stringr::str_extract(.y[1, 2], "(?<=Round:\\s)(.*)(?=\\sVenue)"),
       Venue = stringr::str_extract(.y[1, 2], "(?<=Venue:\\s)(.*)(?=\\Date)"),
-      Attendance = stringr::str_extract(.y[1, 2], "(?<=Attendance:\\s)(.*)"),
-      Umpires = .y[6, 3] %>% dplyr::pull()
+      Attendance = as.integer(stringr::str_extract(.y[1, 2], "(?<=Attendance:\\s)(.*)")),
+      Umpires = .y[6, 3] %>% dplyr::pull(),
+      Date = lubridate::dmy_hm(.data$Date),
+      Local.start.time = as.integer(format(.data$Date, "%H%M")),
+      Date = lubridate::ymd(format(.data$Date, "%Y-%m-%d")),
+      Season = as.integer(lubridate::year(.data$Date))
     ))
 
   games_df <- games_df %>%
-    purrr::map2(.y = details, ~ dplyr::mutate(
+    purrr::map2(.y = game_details, ~ dplyr::mutate(
       .x,
       Home.team = .y[2, 2] %>% dplyr::pull(),
       HQ1 = .y[2, 3] %>% dplyr::pull(),
@@ -105,10 +204,14 @@ scrape_afltables_match <- function(match_urls) {
 
   games_df <-
     games_df %>%
-    purrr::list_rbind()
+    purrr::list_rbind() %>%
+    dplyr::mutate(
+      HQET = dplyr::if_else(stringr::str_detect(.data$HQET, "^\\d+\\.\\d+\\.\\d+$"), .data$HQET, NA_character_),
+      AQET = dplyr::if_else(stringr::str_detect(.data$AQET, "^\\d+\\.\\d+\\.\\d+$"), .data$AQET, NA_character_)
+    )
 
   games_df <- games_df %>%
-    dplyr::mutate(Date = gsub("\\([^]]*", "", .data$Date))
+    dplyr::mutate(Date = as.Date(gsub("\\([^]]*", "", .data$Date)))
 
   # Remove columns with NA and abbreviations
   games_df <- games_df[, !(names(games_df) %in% "NA")]
@@ -120,28 +223,7 @@ scrape_afltables_match <- function(match_urls) {
   # Fix names
   names(games_df) <- make.names(names(games_df))
 
-  # nolint start
-  if ("X." %in% names(games_df)) games_df <- dplyr::rename(games_df, Jumper.No. = "X.")
-  if ("X1." %in% names(games_df)) {
-    games_df <- dplyr::rename(games_df,
-      One.Percenters = "X1."
-    )
-  }
-  if ("X.P" %in% names(games_df)) games_df <- dplyr::rename(games_df, TOG = "X.P")
-  # nolint end
-
-  # change column types
-  games_df <- games_df %>%
-    dplyr::filter(!.data$Player %in% c("Rushed", "Totals", "Opposition")) %>%
-    utils::type.convert(na.strings = "NA", as.is = TRUE)
-
   games_cleaned <- games_df %>%
-    dplyr::mutate(
-      Date = lubridate::dmy_hm(.data$Date),
-      Local.start.time = as.integer(format(.data$Date, "%H%M")),
-      Date = lubridate::ymd(format(.data$Date, "%Y-%m-%d")),
-      Season = as.integer(lubridate::year(.data$Date))
-    ) %>%
     tidyr::separate("Player",
       into = c("Surname", "First.name"), sep = ","
     ) %>%
@@ -156,93 +238,45 @@ scrape_afltables_match <- function(match_urls) {
     dplyr::mutate_at(
       dplyr::vars(dplyr::starts_with("Umpire")),
       stringr::str_replace, " \\(.*\\)", ""
-    ) %>%
-    dplyr::mutate(
-      Substitute = dplyr::case_when(
-        stringr::str_detect(.data$Jumper.No., "\u2191") ~ "On",
-        stringr::str_detect(.data$Jumper.No., "\u2193") ~ "Off",
-        TRUE ~ NA_character_),
-      Jumper.No. = stringr::str_remove_all(.data$Jumper.No., "[\u2191\u2193]") %>% as.character()
-      )
+    )
 
- 
 
-  # Check if extra time happened
-  games_extra_time <- games_cleaned |> 
-    dplyr::select("Surname", "First.name", "Playing.for", "Date", "HQET", "AQET") |> 
-    dplyr::filter(
-      stringr::str_detect("HQET", "[^[:ascii:]]")
-      )
-  
-  # Update score columns based on it ET happened
-  if(nrow(games_extra_time > 0)) {
-    extra_time <- TRUE
-    score_cols <- c("HQ1", "HQ2", "HQ3", "HQ4", "HQET", "AQ1", "AQ2", "AQ3", "AQ4", "AQET")
-  } else {
-    extra_time <- FALSE
-    score_cols <- c("HQ1", "HQ2", "HQ3", "HQ4", "AQ1", "AQ2", "AQ3", "AQ4")
-    
-    games_cleaned <- games_cleaned |> 
-      dplyr::mutate(HQET = NA,
-                    AQET = NA)
-  }
-  
-  
-  # Add ET columns for when it didn't happen
-  
-  games_scores <- games_cleaned |> 
-    tidyr::separate_wider_delim(cols = dplyr::all_of(score_cols),
-                                delim = ".", 
-                                names = c("G", "B", "P"), 
-                                names_sep = "") 
+  # Update score columns
+  score_cols <- c("HQ1", "HQ2", "HQ3", "HQ4", "HQET", "AQ1", "AQ2", "AQ3", "AQ4", "AQET")
 
-  
+  games_scores <- games_cleaned %>%
+    tidyr::separate_wider_delim(
+      cols = dplyr::all_of(score_cols),
+      delim = ".",
+      names = c("G", "B", "P"),
+      names_sep = ""
+    )
+
   games_scores <- games_scores %>%
     dplyr::mutate_at(dplyr::vars(dplyr::contains("HQ")), as.integer) %>%
     dplyr::mutate_at(dplyr::vars(dplyr::contains("AQ")), as.integer)
-  
-  # Get final score, depending on extra time or not
-  if(extra_time) {
-    games_scores <- games_scores |> 
-      dplyr::mutate(Home.score = .data$HQETP,
-                    Away.score = .data$AQETP)
-  } else {
-    games_scores <- games_scores |> 
-      dplyr::mutate(Home.score = .data$HQ4P,
-                    Away.score = .data$AQ4P)
-  }
 
-  # Get player ids
-  games_index <- min(games_scores$Season):max(games_scores$Season)
-  ids <- get_afltables_player_ids(games_index)
+  # Get final score
+  games_scores <- games_scores %>%
+    dplyr::mutate(
+      Home.score = dplyr::coalesce(.data$HQETP, .data$HQ4P),
+      Away.score = dplyr::coalesce(.data$AQETP, .data$AQ4P)
+    )
 
   # Join data
   games_joined <- games_scores %>%
-    dplyr::mutate(
-      Player = paste(.data$First.name, .data$Surname),
-      Team = replace_teams(.data$Playing.for)
-    ) %>%
-    dplyr::left_join(ids %>% dplyr::mutate(Team = replace_teams(.data$Team)),
-      by = c("Season", "Player", "Team")
-    ) # %>% dplyr::select(-"Player")
+    dplyr::mutate(Team = replace_teams(.data$Playing.for)) %>%
+    dplyr::left_join(
+      player_mapping_afltables %>%
+        dplyr::select("url", Player = "player", "ID"),
+      by = c("url")
+    ) 
 
-  df <- games_joined %>%
-    dplyr::rename(!!!rlang::syms(with(stat_abbr, setNames(stat.abb, stat))))
-
-  if (!rlang::has_name(games_joined, "Substitute")) {
-    afldata_cols <- afldata_cols[afldata_cols != "Substitute"]
-  }
-
-  # df <- df %>%
-  #   dplyr::select(dplyr::one_of(afldata_cols))
+  df <- games_joined %>% 
+    dplyr::rename(dplyr::any_of(mapping_afltables))
 
   df <- df %>%
-    dplyr::mutate(dplyr::across(
-      dplyr::where(is.numeric) & !dplyr::any_of(c("HQETG", "HQETB", "HQETP", "AQETG", "AQETB", "AQETP")),
-      ~ ifelse(is.na(.), 0, .)
-    )) %>%
-    dplyr::mutate(Round = as.character(.data$Round))
-
+    dplyr::select(dplyr::any_of(dictionary_afltables$field))
 
   return(df)
 }
@@ -324,6 +358,90 @@ get_afltables_urls <- function(start_date,
   match_ids[!is.na(match_ids)]
 }
 
+#' Convert afltables column formats
+#'
+#' Converts afltables to logical, integer, numeric, character or factor as appropriate
+#'
+#' @param df df to check formatting for
+#' @param column_formats list that contains formatting of potential columns
+#'
+#'
+#' @keywords internal
+#' @noRd
+check_and_convert <- function(df, column_formats) {
+  # Check if column_formats is a dataframe
+  if (!is.data.frame(column_formats)) {
+    stop("column_formats must be a dataframe.")
+  }
+
+  # Check if required columns are in column_formats
+  required_columns <- c("field", "data_type")
+  if (!all(required_columns %in% names(column_formats))) {
+    stop(paste("column_formats must contain the following columns:", paste(required_columns, collapse = ", ")))
+  }
+
+  df <- df %>%
+    dplyr::mutate(dplyr::across(dplyr::any_of(column_formats$field), ~ {
+      data_type <- column_formats$data_type[column_formats$field == dplyr::cur_column()]
+      if (data_type == "character") {
+        as.character(.)
+      } else if (data_type == "numeric") {
+        as.numeric(.)
+      } else if (data_type == "integer") {
+        as.integer(.)
+      } else if (data_type == "date") {
+        as.Date(., format = "%Y-%m-%d")
+      } else if (data_type == "logical") {
+        as.logical(.)
+      } else if (data_type == "factor") {
+        as.factor(.)
+      } else {
+        . # No conversion if data_type is not recognized
+      }
+    }))
+  return(df)
+}
+
+
+
+#' Convert afltables ages into decimal years
+#'
+#' Converts afltables ages into decimal years
+#'
+#' @param age_str age string for conversion
+#'
+#' @keywords internal
+#' @noRd
+# Function to convert age string to decimal years
+convert_age_to_years <- function(age_str) {
+  years <- as.numeric(sub("y.*", "", age_str))
+  days <- as.numeric(sub(".*y ([0-9]+)d.*", "\\1", age_str))
+  decimal_years <- years + (days / 365.25)
+  return(decimal_years)
+}
+
+#' Replace NAs depending on column values
+#'
+#' @param df dataframe for replacing NAs
+#'
+#' @keywords internal
+#' @noRd
+# Function to replace NAs based on the condition
+replace_na_conditionally <- function(df) {
+  df %>% dplyr::mutate(dplyr::across(dplyr::everything(), ~ {
+    if (all(is.na(.))) {
+      .
+    } else {
+      if (is.numeric(.)) {
+        tidyr::replace_na(., 0)
+      } else {
+        tidyr::replace_na(., "") # or some other placeholder value for non-numeric columns
+      }
+    }
+  }))
+}
+
+
 #' Get afltables player ids
 #'
 #' Returns the ID for the players in the relevent seasons
@@ -391,6 +509,7 @@ get_afltables_player_ids <- function(seasons) {
     4643, "Mal Markillie",
     1931, "Mark Maclure", # lower case L
     774, "Mathew Capuano",
+    11746, "Matt de Boer",
     7540, "Morrie Davidson",
     5234, "Percy Blencowe",
     4737, "Roy Wawn",
@@ -502,3 +621,6 @@ get_afltables_player_ids <- function(seasons) {
 
   return(ids)
 }
+
+# silence global variable NOTES
+utils::globalVariables(names = c("dictionary_afltables", "mapping_afltables", "player_mapping_afltables"))