+++ R/helpers-footywire.R
@@ -93,7 +93,7 @@ footywire_html <- function(x, id) {
     rvest::html_text()
 
   # Now get the table data. The Home Team is in the 13th table
-  
+
   home_stats <- x %>%
     rvest::html_nodes("table") %>%
     .[[13]] %>%
@@ -106,9 +106,11 @@ footywire_html <- function(x, id) {
     dplyr::mutate(
       dplyr::across(
         dplyr::where(is.character),
-        ~ dplyr::na_if(.x, "Unused Substitute"))) %>%
+        ~ dplyr::na_if(.x, "Unused Substitute")
+      )
+    ) %>%
     dplyr::mutate(dplyr::across(
-      c(-"Player",-"Team", -"Opposition",-"Status"),
+      c(-"Player", -"Team", -"Opposition", -"Status"),
       as.numeric
     ))
 
@@ -125,9 +127,11 @@ footywire_html <- function(x, id) {
     dplyr::mutate(
       dplyr::across(
         dplyr::where(is.character),
-        ~ dplyr::na_if(.x, "Unused Substitute"))) %>%
+        ~ dplyr::na_if(.x, "Unused Substitute")
+      )
+    ) %>%
     dplyr::mutate(dplyr::across(
-      c(-"Player",-"Team", -"Opposition",-"Status"),
+      c(-"Player", -"Team", -"Opposition", -"Status"),
       as.numeric
     ))