+++ NEWS.md
@@ -1,8 +1,9 @@
 # fitzRoy (development version)
 
 ## New Features
-* Combined scraping functions into a single cohesive function for better efficiency.
-* Include Extra-Time scores for any games that went to extra-time (HQET and AQET)  ([#155](https://github.com/jimmyday12/fitzRoy/issues/155))
+* Include Extra-Time scores in `fetch_player_stats_afltables` for any games that went to extra-time (HQET and AQET)  ([#155](https://github.com/jimmyday12/fitzRoy/issues/155))
+* Include afltables url, Age, and Career.Games as at that game for each player in `fetch_player_stats_afltables`
+* Include Coach and Home.Away for each team for each game in `fetch_player_stats_afltables`
 
 ## Bug Fixes
 * Corrected date parsing issues for AFL tables data, ensuring accurate date formats. ([#136](https://github.com/jimmyday12/fitzRoy/issues/136))